+++ infra/resources/modules/session_manager/iam.tf
@@ -32,6 +32,14 @@ module "ca_iam" {
     }
   }]
 
+  storage_table = [{
+    storage_account_name = var.locked_profiles.storage_account.name
+    resource_group_name  = var.locked_profiles.storage_account.resource_group_name
+    table_name           = var.locked_profiles.table_name
+    role                 = "reader"
+    description          = "Allow Session Manager Container App to read the locked profiles table"
+  }]
+
   # cosmos = [ # TODO: capire su quali DB necessita scrittura/lettura
   #   {
   #     account_name        = data.azurerm_cosmosdb_account.cosmos.name