+++ infra/resources/prod/function_session_manager_internal.tf
@@ -124,7 +124,7 @@ module "function_session_manager_internal_autoscale" {
 
   scheduler = {
     normal_load = {
-      minimum = 10
+      minimum = 3
       default = 10
     },
     maximum = 30