+++ apps/io-session-manager-oi/CHANGELOG.md
@@ -1,5 +1,26 @@
 # io-session-manager-oi
 
+## 0.4.0
+
+### Minor Changes
+
+- cb26d6b: split healthchecks into readiness and liveness probes
+
+## 0.3.0
+
+### Minor Changes
+
+- c683e15: Added sm-internal-rollout port and adapter
+- d352687: Add Locked Profiles outbound adapter
+- a3672f0: Added fn fast-login port and adapter
+
+### Patch Changes
+
+- Updated dependencies [d352687]
+- Updated dependencies [a3672f0]
+  - @pagopa/azure-sdk@0.1.0
+  - @pagopa/io-auth-n-identity-domain@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes